refactor: relocate PiRunner to src/agent/runners/pi/runner.ts (#632 segment 2)
Pure relocate, no logic change. Bottom-up step before pi-mono redesign:
moves the runner to its new home so the upcoming pi-mono → service-cache
+ session-factory + tool-adapter split can target the new location
directly without re-touching imports later.

- src/legacy/agents/runners/pi.ts → src/agent/runners/pi/runner.ts
- src/legacy/agents/runtime.ts: import path updated
- 4 import paths inside the moved file rewired to ../../../legacy/...
  (pi-mono, tools, builtin/system-prompt, agents/types — all still in
  legacy/ for now)

tsc clean; 1158/1158 tests pass.
